Created a test using the golden master principle
The golden master is recorded using the TextTestFixture. I simply
increased the days to 100.

Added a unit test to replay the application and capture the output so
that I can use it to compare against the golden master.

This is the easiest and fastest way to start refactoring with the confidence
that i don't break current functionality.