Added new feature to handle Conjured items
I'm sufficiently happy with the refactoring. The code is readable and
I'm able to implement the new feature without having to worry about
causing side effects.

I'm not 100% I've interpreted the new requirement correctly though, The
conjured items degrade twice as fast, for me that means that past the
sellIn it decreases with 4 points, instead of the normal 2.

If adapted the golden-master.txt file to add test cases for the new
requirement.